Analysis

Documentation

data AnalysisResult

Analyzing a pattern group with a pattern prototype (occ1.csv).

Constructors

An

Fields

total :: Int

all other occurences of the pattern

exact :: Int

# of exact occurences

transposed :: Int

# of transpositions

inverted :: Int

# of inversions

augmented :: Int

# of augmentations

retrograded :: Int

# of retrogrades

rotated :: Int

# of rotations

trInverted :: Int

# of transposed inversions

trAugmented :: Int

# of transposed augmentations

trRetrograded :: Int

# of transposed retrogrades

approxEq2 :: Int

# of approximately exact occurences (dist = 2)

approxEq4 :: Int

# of approximately exact occurences (dist = 4)

approxEq6 :: Int

# of approximately exact occurences (dist = 6)

approxEq8 :: Int

# of approximately exact occurences (dist = 8)

approxEq10 :: Int

# of approximately exact occurences (dist = 10)

approxEq12 :: Int

# of approximately exact occurences (dist = 12)

approxEq14 :: Int

# of approximately exact occurences (dist = 14)

approxEq16 :: Int

# of approximately exact occurences (dist = 16)

analysePatternGroup :: PatternGroup -> AnalysisResult

Analyze a single pattern group

combineAnalyses :: [AnalysisResult] -> AnalysisResult

Combine analyses from different pattern groups.

percentage :: AnalysisResult -> (AnalysisResult -> Int) -> Double

Get the percentage of an equivalence class from an analysis result.

getTitle :: PatternGroup -> String

Get a title, unique to the given PatternGroup, in the following format: piece:expert:pattern.