Charts

Documentation

renderOne :: PatternGroup -> AnalysisResult -> IO ()

Visualize the results of analyzing a single pattern group in a pie chart.

renderAll :: String -> AnalysisResult -> IO ()

Visualize the result of analyzing multiple pattern groups in a single pie chart.