Discovery

Documentation

type WindowSize = Int

type Query a = (Check a, a)

data UserQuery a

Constructors

ToPattern a => (Check Pattern) :@ a

upTo :: Time -> Time -> (Time, Time)

query :: Query Pattern -> MusicPiece -> [Pattern]

Query equivalent patterns using a sliding window.

query1 :: UserQuery (Time, Time)

Example queries.

query2 :: UserQuery (Music Pitch)

(??) :: ToPattern a => Song -> UserQuery a -> IO () infix 0

Query patterns from the given song with given base pattern. e.g. "bach" ?? query1/query2

class ToPattern a where

Types from which we can extract a pattern from a given song.

Minimal complete definition

toPattern

Methods

toPattern :: MusicPiece -> a -> Pattern