Transformations

Documentation

newtype Check a

Constructors

Check
Checks two patterns (horizontal translation in time is always assumed).

Fields

getCheck :: a -> a -> Bool

(<=>) :: a -> a -> Check a -> Bool

(>$<) :: (a -> b) -> Check b -> Check a infix 7

(>$) :: (a -> a) -> Check a -> Check a infix 8

($<) :: (a -> a) -> Check a -> Check a infix 8

exactOf :: Check Pattern

Exact repetition: move a pattern in time. (AKA horizontal translation)

transpositionOf :: Check Pattern

Transposition: move a pattern in pitch. (AKA horizontal+vertical translation)

inversionOf :: Check Pattern

Inversion: negate all pitch intervals (starting from the same base pitch).

retrogradeOf :: Check Pattern

Retrograde: mirror a pattern in both pitch and rhythm. (AKA vertical reflection)

rotationOf :: Check Pattern

Rotation: reversal of pitch intervals and reversal of rhythm. (AKA retrograde inversion)

augmentationOf :: Check Pattern

Augmentation: speed-up/slow-down the rhythmic structure of a pattern.

type Degree = Integer

type Scale = Map MIDI Degree

tonalTranspOf :: Check Pattern

Octave-agnostic tonal transposition, wrt a scale that fits the base pattern e.g. [I, IV, V] tonalTranspOf [III, VI, VII]

trInversionOf :: Check Pattern

Transposition + Inversion. (AKA horizontal reflection)

trAugmentationOf :: Check Pattern

Transposition + Augmentation.

trRetrogradeOf :: Check Pattern

Transposition + Retrograde. (AKA vertical glide reflection)

type Interval = Integer

equal :: Eq a => Check a

Check that two elements are exactly equal (using eq). e.g. [a, c, b] equal [a, c, b]

approxEq :: Float -> Check Pattern

Check that two patterns are approximately equal, wrt a certain percentage. A base pattern and an occurence are approximately equal with percentage p when: 1. The occurence ignores (1-p)% notes of the base pattern 2. (1-p)% notes of the occurence are additional notes (not in the base pattern) e.g. [A,C,F,A,B] (approxEq 80%) [A,C,G,A,B]

inverse :: Num a => [a] -> [a]

Negate the values of a numeric list.

basePitch :: Pattern -> Maybe MIDI

The base pitch of a pattern (the pitch of its first note). e.g. basePitch [(25,1), (27,2), (25,2.5)] = Just 25

pitch :: Pattern -> [MIDI]

The (real) pitch structure of a pattern. e.g. pitch [(25,1), (27,2), (25,2.5)] = [25, 27, 25]

intervals :: Pattern -> [Interval]

The (relative) pitch structure of a pattern. e.g. intervals [(25,1), (27,2), (25,2.5)] = [2, -2]

durations :: Pattern -> [Time]

The (real) rhythmic structure of a pattern. e.g. durations [(25,1), (27,2), (25,2.5)] = [1, 2, 2.5]

rhythm :: Pattern -> [Time]

The (relative) rhythmic structure of a pattern. e.g. rhythm [(25,1), (27,2), (25,2.5)] = [1, 1.5]

normalizeTime :: [Time] -> [Time]

Convert times to ratios wrt the first time unit used. e.g. normalizeTime [2, 6, 8, 6, 1, 2] = [1, 3, 4, 1/2, 1]

normalRhythm :: Pattern -> [Time]

Normalized (relative) rhythmic structure of a pattern. e.g. normalRhythm [(A,2), (C#,6), (Eb,8), (B,1), (A,2)] = [1, 3, 4, 1/2, 1]

translateH :: Time -> Note -> Note

Translate a note horizontally (in time). e.g. translateH (-0.5) [(25,1), (27,2), (25,2.5)] = [(25,0.5), (27,1.5), (25,2)]

translateV :: Interval -> Note -> Note

Translate a note vertically (in pitch). e.g. translateV (-20) [(25,1), (27,2), (25,2.5)] = [(5,1), (7,2), (5,2.5)]

pairs :: [a] -> [(a, a)]

Get list as pairs of consecutive elements. e.g. pairs [a, b, c, d] = [(a, b), (b, c), (c, d)]