Types

Documentation

type Time = Double

Time in crochet beats.

type MIDI = Integer

MIDI values are represented with integers.

type Interval = Integer

Intervals are represented with integers (i.e. number of semitones).

data PatternGroup

A pattern group is one of the patterns of a piece of music, identified by an expert or algorithm, and defined by a pattern prototype and other pattern occurences.

Constructors

PatternGroup

Fields

piece_name :: String

the name of the music piece, that the pattern group belongs to

expert_name :: String

music expert or algorithm that produced the pattern occurences

pattern_name :: String

the name of the current pattern group

basePattern :: Pattern

the pattern prototype (always taken from `occ1.csv`)

patterns :: [Pattern]

all other pattern occurences of the prototype

type Pattern = [Note]

A pattern is a sequence of notes.

data Note

A simplistic music note (only time and pitch).

Constructors

Note

Fields

ontime :: Time

onset time

midi :: MIDI

MIDI number

(.@) :: Time -> MIDI -> Note

Infix variant of the Note constructor.

type MusicPiece = Pattern

A piece of music is a huge pattern.

type Song = String

Songs are identified with a string.

pmap :: (a -> b) -> [a] -> [b]

Parallel map.

pforM :: Traversable t => t a -> (a -> IO b) -> IO (t b)

Parallel forM.

pmapM :: Traversable t => (a -> IO b) -> t a -> IO (t b)