Analysis

Documentation

data AnalysisResult

Analyzing a pattern group with a pattern prototype (occ1.csv).

Constructors

An

Fields

name :: !String

name associated with the analysis

total :: !Int

all other occurences of the pattern

exact :: !Int

# of exact occurences

transposed :: !Int

# of *atonal* transpositions

tonalTransped :: !Int

# of *tonal* transpositions

inverted :: !Int

# of inversions

augmented :: !Int

# of augmentations

retrograded :: !Int

# of retrogrades

rotated :: !Int

# of rotations

trInverted :: !Int

# of transposed inversions

trAugmented :: !Int

# of transposed augmentations

trRetrograded :: !Int

# of transposed retrogrades

exact8 :: !Int

# of exact occurences

transposed8 :: !Int

# of *atonal* transpositions

tonalTransped8 :: !Int

# of *tonal* transpositions

inverted8 :: !Int

# of inversions

augmented8 :: !Int

# of augmentations

retrograded8 :: !Int

# of retrogrades

rotated8 :: !Int

# of rotations

trInverted8 :: !Int

# of transposed inversions

trAugmented8 :: !Int

# of transposed augmentations

trRetrograded8 :: !Int

# of transposed retrogrades

exact6 :: !Int

# of exact occurences

transposed6 :: !Int

# of *atonal* transpositions

tonalTransped6 :: !Int

# of *tonal* transpositions

inverted6 :: !Int

# of inversions

augmented6 :: !Int

# of augmentations

retrograded6 :: !Int

# of retrogrades

rotated6 :: !Int

# of rotations

trInverted6 :: !Int

# of transposed inversions

trAugmented6 :: !Int

# of transposed augmentations

trRetrograded6 :: !Int

# of transposed retrogrades

exact4 :: !Int

# of exact occurences

transposed4 :: !Int

# of *atonal* transpositions

tonalTransped4 :: !Int

# of *tonal* transpositions

inverted4 :: !Int

# of inversions

augmented4 :: !Int

# of augmentations

retrograded4 :: !Int

# of retrogrades

rotated4 :: !Int

# of rotations

trInverted4 :: !Int

# of transposed inversions

trAugmented4 :: !Int

# of transposed augmentations

trRetrograded4 :: !Int

# of transposed retrogrades

exact2 :: !Int

# of exact occurences

transposed2 :: !Int

# of *atonal* transpositions

tonalTransped2 :: !Int

# of *tonal* transpositions

inverted2 :: !Int

# of inversions

augmented2 :: !Int

# of augmentations

retrograded2 :: !Int

# of retrogrades

rotated2 :: !Int

# of rotations

trInverted2 :: !Int

# of transposed inversions

trAugmented2 :: !Int

# of transposed augmentations

trRetrograded2 :: !Int

# of transposed retrogrades

unclassified :: ![(String, Pattern)]

filenames of all unclassified patterns

defAn :: AnalysisResult

singleAn :: AnalysisResult

analysePatternGroup :: Bool -> PatternGroup -> IO AnalysisResult

Analyze a single pattern group.

combineAnalyses :: [AnalysisResult] -> AnalysisResult

Combine analyses from different pattern groups.

percentage :: AnalysisResult -> (AnalysisResult -> Int) -> Double

Get the percentage of an equivalence class from an analysis result.

otherPercentage :: AnalysisResult -> Double

Get the percentage of unclassified patterns from an analysis result.

Orphan instances

ToField [(String, Pattern)]

Methods

toField :: [(String, Pattern)] -> Field