Types

Documentation

type Time = Double

Time in crochet beats.

type MIDI = Integer

MIDI values are represented with integers.

type MusicPiece = [MirexEntry]

A piece of music consists of multiple music events.

data MirexEntry

A single entry in a MIREX piece of music.

Constructors

MirexEntry

Fields

t :: Time

onset time

mid :: MIDI

note's MIDI number

morph :: Integer

note's morphetic pitch number

duration :: Time

duration in crotchet beats

staff :: Integer

voice number (starting from 0)

data PatternGroup

A pattern group is one of the patterns of a piece of music, identified by an expert or algorithm, and defined by a pattern prototype and other pattern occurences.

Constructors

PatternGroup

Fields

piece_name :: String

the name of the music piece, that the pattern group belongs to

expert_name :: String

music expert or algorithm that produced the pattern occurences

pattern_name :: String

the name of the current pattern group

basePattern :: Pattern

the pattern prototype (always taken from `occ1.csv`)

patterns :: [Pattern]

all other pattern occurences of the prototype

type Pattern = [Note]

A pattern is a sequence of notes.

data Note

A simplistic music note (only time and pitch).

Constructors

Note

Fields

ontime :: Time

onset time

midi :: MIDI

MIDI number

newtype Check a

Constructors

Check
Checks two patterns (horizontal translation in time is always assumed).

Fields

getCheck :: a -> a -> Bool

(<=>) :: a -> a -> Check a -> Bool

(>$<) :: (a -> b) -> Check b -> Check a infix 7

(>$) :: (a -> a) -> Check a -> Check a infix 8

($<) :: (a -> a) -> Check a -> Check a infix 8

exactOf :: Check Pattern

Exact repetition: move a pattern in time. (AKA horizontal translation)

transpositionOf :: Check Pattern

Transposition: move a pattern in pitch. (AKA horizontal+vertical translation)

inversionOf :: Check Pattern

Inversion: negate all pitch intervals (starting from the same base pitch).

retrogradeOf :: Check Pattern

Retrograde: mirror a pattern in both pitch and rhythm. (AKA vertical reflection)

rotationOf :: Check Pattern

Rotation: reversal of pitch intervals and reversal of rhythm. (AKA retrograde inversion)

augmentationOf :: Check Pattern

Augmentation: speed-up/slow-down the rhythmic structure of a pattern.

trInversionOf :: Check Pattern

Transposition + Inversion. (AKA horizontal reflection)

trAugmentationOf :: Check Pattern

Transposition + Augmentation.

trRetrogradeOf :: Check Pattern

Transposition + Retrograde. (AKA vertical glide reflection)

type Interval = Integer

equal :: Eq a => Check a

Check that two elements are exactly equal (using eq). e.g. [a, c, b] equal [a, c, b]

inverse :: Num a => [a] -> [a]

Negate the values of a numeric list.

basePitch :: Pattern -> Maybe MIDI

The base pitch of a pattern (the pitch of its first note). e.g. basePitch [(25,1), (27,2), (25,2.5)] = Just 25

pitch :: Pattern -> [MIDI]

The (real) pitch structure of a pattern. e.g. pitch [(25,1), (27,2), (25,2.5)] = [25, 27, 25]

intervals :: Pattern -> [Interval]

The (relative) pitch structure of a pattern. e.g. intervals [(25,1), (27,2), (25,2.5)] = [2, -2]

durations :: Pattern -> [Time]

The (real) rhythmic structure of a pattern. e.g. durations [(25,1), (27,2), (25,2.5)] = [1, 2, 2.5]

rhythm :: Pattern -> [Time]

The (relative) rhythmic structure of a pattern. e.g. rhythm [(25,1), (27,2), (25,2.5)] = [1, 1.5]

normalizeTime :: [Time] -> [Time]

Convert times to ratios wrt the first time unit used. e.g. normalizeTime [2, 6, 8, 6, 1, 2] = [1, 3, 4, 1/2, 1]

normalRhythm :: Pattern -> [Time]

Normalized (relative) rhythmic structure of a pattern. e.g. normalRhythm [(A,2), (C#,6), (Eb,8), (B,1), (A,2)] = [1, 3, 4, 1/2, 1]

translateH :: Time -> Note -> Note

Translate a note horizontally (in time). e.g. translateH (-0.5) [(25,1), (27,2), (25,2.5)] = [(25,0.5), (27,1.5), (25,2)]

translateV :: Interval -> Note -> Note

Translate a note vertically (in pitch). e.g. translateV (-20) [(25,1), (27,2), (25,2.5)] = [(5,1), (7,2), (5,2.5)]

pairs :: [a] -> [(a, a)]

Get list as pairs of consecutive elements. e.g. pairs [a, b, c, d] = [(a, b), (b, c), (c, d)]